requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
aiostream=0.5.2=pypi_0
alembic=1.13.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
blas=1.0=mkl
boto3=1.34.144=pypi_0
botocore=1.34.144=pypi_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorlog=6.8.2=pypi_0
curl=8.7.1=hdbd6064_0
datasets=2.20.0=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
dspy-ai=2.4.10=pypi_0
email-validator=2.2.0=pypi_0
expat=2.6.2=h6a678d5_0
faiss-cpu=1.8.0.post1=pypi_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
filelock=3.15.4=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
func-timeout=4.3.5=pypi_0
gdbm=1.18=hd4cb3f1_4
gettext=0.21.0=h39681ba_1
git=2.45.2=pl5340h9abc3c3_0
greenlet=3.0.3=pypi_0
grpclib=0.4.7=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.4=pypi_0
hyperframe=6.0.1=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libcurl=8.7.1=h251f7ec_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxml2=2.10.4=hfdd30dd_2
mako=1.3.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openai=1.34.0=pypi_0
openssl=3.0.14=h5eee18b_0
optuna=3.6.1=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pcre2=10.42=hebb0a14_1
perl=5.34.0=h5eee18b_2
pillow=10.4.0=pypi_0
pip=24.0=py311h06a4308_0
protobuf=4.25.3=pypi_0
psycopg2-binary=2.9.9=pypi_0
pyarrow=16.1.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11-abi=4=hd3eb1b0_1
pydantic=2.7.4=pypi_0
pydantic-core=2.18.4=pypi_0
pygments=2.18.0=pypi_0
pysqlite3=0.5.3=pypi_0
python=3.11.5=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
s3transfer=0.10.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.4.2=py311ha02d727_1
scipy=1.13.1=py311h08b1b3b_0
sentence-transformers=3.0.1=pypi_0
setuptools=69.5.1=py311h06a4308_0
shellingham=1.5.4=pypi_0
sigtools=4.0.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.31=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.37.2=pypi_0
structlog=24.2.0=pypi_0
sympy=1.13.0=pypi_0
synchronicity=0.6.7=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py311h92b7b1e_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
toml=0.10.2=pypi_0
torch=2.3.1=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.42.4=pypi_0
triton=2.3.1=pypi_0
typer=0.12.3=pypi_0
types-certifi=2021.10.8.3=pypi_0
types-toml=0.10.8.20240310=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.3=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py311h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_1
